requirements.txt
matplotlib==3.3.4
numpy==1.20.1
openai==0.27.6
pandas==1.2.4
Pillow==8.3.1
scipy==1.6.2
streamlit==1.11.0
timm==0.6.13
torch==1.9.0
torchvision==0.10.0
altair == 4.2.2
